% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_datasets.R
\name{explore_datasets}
\alias{explore_datasets}
\title{Explore Datasets}
\usage{
explore_datasets(
  entries,
  format = "/?res_format=CSV&res_format=JSON",
  OrgType = NA
)
}
\arguments{
\item{entries}{the amount of data set names to return as an integer}

\item{format}{What data format you want to download. Default you look for both csv and json, as that is what get_gov_file currently supports. You can use csv (only csv files)  or json (only json files). You can use parameter "all" to see all
datasets no matter what format it is.}

\item{OrgType}{What organization type published the data. Options are  "City Government", "State Government","Federal Government" , "County Government", "University"}
}
\value{
a character vector with the names of datasets
}
\description{
Returns to  users a desired amount of dataset names they could download. They can filter the datasets by format or Organization type. Meant to be used with get_gov_file function.
}
\examples{
library(CodeGovern)
# 2 Parameters, 2 pages
titles1 <- explore_datasets(25, format = "csv", OrgType = "Federal Government")
# no parameters, 1 page scraped
titles2 <- explore_datasets(10)
}
