% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federal_alert.R
\name{federal_alert}
\alias{federal_alert}
\title{Alert user if data is Non-Federal}
\usage{
federal_alert(name)
}
\arguments{
\item{name}{name of the data set}
}
\value{
A statement notifying whether the data is Federal or Non-Federal
}
\description{
Alert user if data is Non-Federal
}
\examples{
federal_alert("Electric Vehicle Population Data")
federal_alert("FDIC Failed Bank List")
}
