% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federal_alert.R
\name{federalalert}
\alias{federalalert}
\title{Alert user if data is Non-Federal}
\usage{
federalalert(name)
}
\arguments{
\item{name}{name of the data set}
}
\value{
A statement notifying whether the data is Federal or Non-Federal
}
\description{
Alert user if data is Non-Federal
}
\examples{
federalalert("Electric Vehicle Population Data")
federalalert("FDIC Failed Bank List")
}
