% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gov_file.R
\name{get_gov_file}
\alias{get_gov_file}
\title{Download/Import CSV, JSON, XML files.}
\usage{
get_gov_file(name, type, import = TRUE)
}
\arguments{
\item{name}{name of the data set}

\item{type}{data file type (csv, json, xml)}

\item{import}{Decides if we locally download the file. By default import is TRUE.}
}
\value{
Locally download the data file and import it to the RStudio environment
}
\description{
Scrape \url{http://www.data.gov} website and automatically download data.  For CSV file type, the function will download the file locally and return a data frame.
For JSON the function will download the file locally and return the raw JSON file. For XML, the function will download the file locally.
}
\examples{

library(CodeGovern)

get_gov_file("Electric Vehicle Population Data", "json", import = FALSE )
get_gov_file("Lottery Powerball Winning Numbers Beginning 2010", "csv", import = FALSE)
}
