% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_many_gov_files.R
\name{get_many_gov_files}
\alias{get_many_gov_files}
\title{Download and Import Multiple CSV Data Sets at the Same Time}
\usage{
get_many_gov_files(name_vector, import = TRUE)
}
\arguments{
\item{name_vector}{The vector of the names of the datasets to be downloaded.}

\item{import}{Decides if we locally download the file. By default import is TRUE.}
}
\value{
Locally download the data files and import them to the RStudio environment as a list of dataframes.
}
\description{
Scrape \url{http://www.data.gov} website and automatically download CSV type data. The function takes a vector of data set names, downloads each data set, and imports
it into the RStudio environment as a list of dataframes.
}
\examples{

library(CodeGovern)

#if you want file to locally import, delete import parameter, by default import = TRUE
my_names <- c("FDIC Failed Bank List", "Demographic Statistics By Zip Code")
get_many_gov_files(my_names, import = TRUE)
}
