% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{Lottery2010}
\alias{Lottery2010}
\title{Lottery Powerball Winning Numbers Beginning 2010}
\format{
\subsection{\code{Lottery2010}}{

A data frame with 1,390 rows and 3 columns:
\describe{
\item{draw_date}{date the winning numbers were drawn}
\item{Winning_numbers}{The 6 winning numbers separated by a space}
\item{Multiplier}{The multiplier number drawn}
...
}
}
}
\source{
\url{https://catalog.data.gov/dataset/lottery-powerball-winning-numbers-beginning-2010}
}
\usage{
Lottery2010
}
\description{
Powerball lottery winning numbers released by the state of New York City
and maintained by Open Data NY,
}
\keyword{datasets}
