% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gov_file.R
\name{get_gov_file}
\alias{get_gov_file}
\title{CodeGovt}
\usage{
get_gov_file(name, type)
}
\arguments{
\item{name}{name of the data set}

\item{type}{data file type (csv, json, xml)}
}
\value{
Locally download the data file and import it to the RStudio environment
}
\description{
Scrape data.gov website and automatically download data.  For CSV file type, the function will download the file locally and return a data frame.
For JSON the function will download the file locally and return the raw JSON file. For XML, the function will download the file locally.
}
\examples{
get_gov_file("Electric Vehicle Population Data", "json")
get_gov_file("Lottery Powerball Winning Numbers Beginning 2010", "csv")
}
